dh_installsystemd: add --reload-after-upgrade option
which uses systemd's reload functionality to reload units instead of
restarting, if supported by the unit.

when used in combination with --restart-after-upgrade, systemd's
(try-)reload-or-restart action is used, to allow utilization for both
the "reload if running" and "reload if possible, start if not running"
use cases.

this functionality is intended for services which support re-executing
themselves via ExecReload without dropping existing connections/clients.

requires support by deb-systemd-invoke[1,2], else the reload action will
not be forwarded to systemctl correctly.

1: https://salsa.debian.org/debian/init-system-helpers/-/merge_requests/5
2: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=892958

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>