Dh_Lib: Add an error handler to variable_substitution to ignore errors
Make possible to ignore value sobstitution errors when a substitution
failed, in this way we can make it ignore changing variables that it
doesn't handle natively.

For example it can be used to parse a debian/control.in file that might
contain variables in the form ${misc:Depends} but we don't want it to
handle those, so instead of using ${Dollar}{...} we can just use
variable substitution with relaxed checks.

To make this working we unfortunately have to use a workaround to make
the substitution work when we skip a value, temporary replacing it with
a stub version of it (${a:b} becomes $~~{~~a:b~~}~~) that are eventually
replaced to the fixed version.