dh_listpackages: Package-Type: udeb implies Build-Profiles: <!noudeb>
Many packages that build udebs, can do so with pure dh sequencer
alone. Many udeb packages only declare package type udeb, and do not
declare any build-profiles. This makes them not support noudeb build
profile. However, it would be really nice if they did. To speed up
dh_builddeb / upload especialy when udebs are discarded and not needed
in the end. Also doing this in debhelper will save everyone sprinkling
those all over the archive (i.e. the dozens and dozens of d-i packages
components).

This may cause a small minority of packages to FTBFS, thus adding this
in the next compat level might be best.

Also not sure if this needs Policy wording changes too or not.